requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pandas==2.3.2
openpyxl==3.1.5
qrcode==8.2
Pillow==11.3.0
gunicorn==23.0.0
email-validator==2.2.0
